Create 45-Minute-Code-1
During High School my teacher would occasionally have a "pop coding" day over topics that we had learned in class. This is an example of one of those projects. The objective of this program is to identify the smallest and largest values in a set of inputted values, and determine which values are the remainder.