Create Random Number Generator
Finds a random number in a given range. In the example used in the code, the user is generating random numbers that simulate rolling a die. The program also asks the user to input a value that they are looking for and calculates the probability of rolling the inputted value for 80 trials.