Create Student Class, Course Class and GPA Calculator
This program creates the classes student and course, which allow a user to input information about the student and course respectively. Then it allows a user to add course information to a student and their schedule. Based on the inputted grades and credit hours, the program will calculate the GPA of a given student.